refactor: restructuring enable_incomplete_progress_transmission to moodle config level
