+++ tests/test_integrated_channels/test_moodle/test_transmitters/test_learner_data.py
@@ -4,11 +4,15 @@ Tests for Moodle learner data transmissions.
 import datetime
 import unittest
 from unittest import mock
+from unittest.mock import Mock
 
 from pytest import mark
 
+from integrated_channels.integrated_channel.exporters.learner_data import LearnerExporter, LearnerExporterUtility
+from integrated_channels.integrated_channel.transmitters.learner_data import LearnerTransmitter
 from integrated_channels.moodle.models import MoodleLearnerDataTransmissionAudit
 from integrated_channels.moodle.transmitters import learner_data
+from integrated_channels.utils import encode_data_for_logging, generate_formatted_log
 from test_utils import factories
 
 
@@ -58,6 +62,8 @@ class TestMoodleLearnerDataTransmitter(unittest.TestCase):
         self.create_course_completion_mock = create_course_completion_mock.start()
         self.addCleanup(create_course_completion_mock.stop)
 
+        self.learner_transmitter = LearnerTransmitter(self.enterprise_config)
+
     def test_transmit_success(self):
         """
         Learner data transmission is successful and the payload is saved with the appropriate data.
@@ -70,3 +76,51 @@ class TestMoodleLearnerDataTransmitter(unittest.TestCase):
         self.create_course_completion_mock.assert_called_with(self.payload.moodle_user_email, self.payload.serialize())
         assert self.payload.status == '200'
         assert self.payload.error_message == ''
+
+    @mock.patch("integrated_channels.integrated_channel.models.LearnerDataTransmissionAudit")
+    @mock.patch('integrated_channels.integrated_channel.transmitters.'
+                'learner_data.LOGGER', autospec=True)
+    def test_incomplete_progress_learner_data_transmission(self, mock_logger, learner_data_transmission_audit_mock):
+        """
+        Test that a customer's configuration can run in enable incomplete progress transmission mode
+        """
+        # Set boolean flag to true
+        self.enterprise_config.enable_incomplete_progress_transmission = True
+
+        self.learner_transmitter.client.create_course_completion = Mock(return_value=(200, 'success'))
+
+        LearnerExporterMock = LearnerExporter
+
+        learner_data_transmission_audit_mock.serialize = Mock(return_value='serialized data')
+        learner_data_transmission_audit_mock.user_id = 1
+        learner_data_transmission_audit_mock.enterprise_course_enrollment_id = 1
+        learner_data_transmission_audit_mock.course_completed = False
+        learner_data_transmission_audit_mock.course_id = 'course_id'
+        LearnerExporterMock.export = Mock(return_value=[learner_data_transmission_audit_mock])
+        lms_user_id = LearnerExporterUtility.lms_user_id_for_ent_course_enrollment_id(
+            learner_data_transmission_audit_mock.enterprise_course_enrollment_id
+        )
+        serialized_payload = learner_data_transmission_audit_mock.serialize(enterprise_configuration=self.enterprise_config)
+        encoded_serialized_payload = encode_data_for_logging(serialized_payload)
+        self.learner_transmitter.transmit(
+            LearnerExporterMock,
+            remote_user_id='user_id'
+        )
+        # with enable_incomplete_progress_transmission = True we should be able to call this method
+        assert self.learner_transmitter.client.create_course_completion.called
+
+        # Set boolean flag to false
+        self.enterprise_config.enable_incomplete_progress_transmission = False
+        self.learner_transmitter.transmit(
+            LearnerExporterMock,
+            remote_user_id='user_id'
+        )
+        mock_logger.info.assert_called_with(generate_formatted_log(
+            self.enterprise_config.channel_code(),
+            self.enterprise_config.enterprise_customer.uuid or None,
+            lms_user_id,
+            learner_data_transmission_audit_mock.course_id,
+            'Skipping in-progress enterprise enrollment record '
+            f'integrated_channel_remote_user_id={learner_data_transmission_audit_mock.user_id}, '
+            f'integrated_channel_serialized_payload_base64={encoded_serialized_payload}'
+        ))
